Add Code Monkey as a Next.js demo
This is an example project to demonstrate how to connect a Next.js project to the Soul Engine.

This soul can help to quickly create a draft version of a code snippet, technical questions, or just chit-chat. Code Monkey is designed to not make any assumptions about coding requests and will ask for clarifications when necessary after writing a rough outline of the code.