Minor correction of typos, and clarification around dispatches (#56)
* Update core.mdx

* Update perceptionAndInteractionRequest.mdx

* Update perceptionAndInteractionRequest.mdx

* Update perceptionAndInteractionRequest.mdx

Wrote a reminder about extracting `dispatch` for Cursor IDE and other new initiates to the Soul Engine :)

* Update perceptionAndInteractionRequest.mdx

I linked out to the usePerceptions page in the wiki, in case a first-time reader hadn't encountered this before. 

I also like the food critic soul as an example of how to make use of usePerceptions for incoming custom actions. 

https://github.com/opensouls/community/blob/60bb932fe34ba765431ac7a6ce83930013732781/souls/food-critic/soul/initialProcess.ts#L9

* Update docs/pages/core.mdx

Co-authored-by: Topper Bowers <topper@quorumcontrol.com>

* Update docs/pages/souls/perceptionAndInteractionRequest.mdx

Co-authored-by: Topper Bowers <topper@quorumcontrol.com>

---------

Co-authored-by: Topper Bowers <topper@quorumcontrol.com>